requirements.txt
# Prod
jsonify>=0.5
flask>=3
requests>=2.31
python-dotenv>=1.0.0
gunicorn>=21.2

# Dev
ipdb
